cleanup(services): move more platform services under lib
