+++ lib/insights.rb
@@ -3,6 +3,8 @@
 require 'insights/api/common/adjust_tags/middleware'
 require 'insights/api/common/satellite_compensation/middleware'
 require 'insights/api/common/logger_with_audit'
+require 'insights/api/common/platform'
+require 'insights/api/common/host_inventory'
 
 # Module to hold every platform-related code
 module Insigths
